refactor: add dto models, create global util file, start yearlyjs
Add DTOs for Day, Month, and Year goals along with goal object (I did
not see that we had decided to scrap the idea of individual goal
objects, so I will have to refactor this in the future).
Add Global Utility file to house functions and objects that will be used
around the repository (saw that Justin may be working on this too. Will
converse later)